The story of our fight with the mighty dark and light mode
After finding my new companion "dati" on my journeys (I found him in an abandoned mine), we headed out to fight the DarkMode boss, he was very hard, but after spending hours fighting him, we didn't even scratch him AT ALL! So we had to use our ultimate weapon the staff of Youtube tutorials, after using the staff, the boss was forced to watch a 6 minute long tutorial on how to swap styles, after whitch he was easy to beat. My new trusty companion dati SLASHED his arm off! He started screaming like a mad man, afterwards I sliced his head off myself, and took it to the nearest village to sell it and buy myself and dati a new meal and horses. The village was quite happy we did so, and even built a statue of us (it was difficult because im tall and dati is veeeery small). But our adventures didn't end there! To be continued